重构OCR修复
