+++ web/src/pages/workspace/Toolbox.vue
@@ -7,16 +7,14 @@ import {
 import { UploadCustomRequestOptions } from 'naive-ui';
 
 import { Locator } from '@/data';
-import { Epub, parseFile, Srt, Txt } from '@/util/file';
+import { ParsedFile, parseFile } from '@/util/file';
 import { downloadFile, downloadFilesPacked } from '@/util';
 
 import { Toolbox } from './Toolbox';
 
 const message = useMessage();
 
-type ToolboxFile = Epub | Txt | Srt;
-
-const files = ref<ToolboxFile[]>([]);
+const files = ref<ParsedFile[]>([]);
 
 const loadFile = async (file: File) => {
   if (files.value.find((it) => it.name === file.name) !== undefined) {
@@ -63,8 +61,6 @@ const customRequest = ({
 };
 
 const showListModal = ref(false);
-
-const fixOcr = () => Toolbox.fixOcr(files.value);
 
 const convertToTxt = () => (files.value = Toolbox.convertToTxt(files.value));
 
@@ -123,17 +119,11 @@ const download = async () => {
 
     <n-list bordered style="margin-top: 20px">
       <n-list-item>
-        <toolbox-item-compress-image />
+        <toolbox-item-compress-image :files="files" />
       </n-list-item>
 
       <n-list-item>
-        <n-flex vertical>
-          <b>TXT：修复OCR换行</b>
-          OCR输出的文本通常存在额外的换行符，导致翻译器错误。当前修复方法是检测每一行的结尾是否是字符（汉字/日文假名/韩文字符/英文字母），如果是的话则删除行尾的换行符。
-          <n-flex>
-            <c-button label="修复" size="small" @action="fixOcr" />
-          </n-flex>
-        </n-flex>
+        <toolbox-item-fix-ocr :files="files" />
       </n-list-item>
 
       <n-list-item>
